refactor!: rename zeitgeist module to temporal
BREAKING CHANGE: Adjective usage is more proper than noun usage, like
that we are using numeric for the module that deals with numbers and
textual for the module that deals with texts.
