+++ src/index.ts
@@ -6,7 +6,7 @@ export * from './-numeric';
 export * as numeric from './-numeric';
 export * from './-purify';
 export * as purify from './-purify';
+export * from './-temporal';
+export * as temporal from './-temporal';
 export * from './-textual';
 export * as textual from './-textual';
-export * from './-zeitgeist';
-export * as zeitgeist from './-zeitgeist';
